Implement menu flyout
